# Draft: jody: fix checkptr segfault