Fix checkptr segfault by disallowing advancing pointer beyond the Go allocation.

Signed-off-by: Yerden Zhumabekov <yerden.zhumabekov@gmail.com>